Database is still not completely working but this is what I have so far. I can't seem to find what's wrong but I'll try to find it tomorrow!
