+++ app/src/main/java/com/example/vetapp/MainActivity.java
@@ -3,12 +3,35 @@ package com.example.vetapp;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+import android.view.View;
+import android.widget.TextView;
+import android.widget.Toast;
 
 public class MainActivity extends AppCompatActivity {
+
+    TextView insert;
+    DBHelper DB = new DBHelper(MainActivity.this);
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        insert = findViewById(R.id.txt_createAccountLink);
+
+        insert.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View view){
+
+
+                boolean checkInsertData = DB.insertUserAccount("user@gmail.com", "password");
+                if(checkInsertData){
+                    Toast.makeText(MainActivity.this, "New Entry Inserted", Toast.LENGTH_SHORT).show();
+                }
+                else{
+                    Toast.makeText(MainActivity.this, "New Entry Not Inserted", Toast.LENGTH_SHORT).show();
+                }
+            }
+        });
     }
 }
