+++ app/src/main/res/layout/activity_main.xml
@@ -123,6 +123,7 @@
         android:layout_height="wrap_content"
         android:text="@string/txt_createAccountLink"
         android:textColor="#2E8BB3"
+        android:textIsSelectable="true"
         android:textStyle="bold"
         app:layout_constraintBottom_toTopOf="@+id/txt_login_username"
         app:layout_constraintEnd_toEndOf="parent"
